kmitl_comp/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.7
django-allauth==0.51.0
django-cors-headers==3.13.0
django-oauth-toolkit==2.1.0
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
jwcrypto==1.4.2
mssql-django==1.1.3
mysqlclient==2.1.1
oauthlib==3.2.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyodbc==4.0.34
pyparsing==3.0.9
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.2.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
singleton-decorator==1.0.0
six==1.16.0
social-auth-core==4.3.0
sqlparse==0.4.2
types-cryptography==3.3.23
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
wincertstore==0.2
wrapt==1.14.1
